wip - add scouting manager and sector-based scouting tracking
